Relocate metrics tracker to shared reporting utilities
The tracker previously lived in src/plugin/src/reporting/metrics.js but is used by shared tests and docs, so move it to src/shared/reporting/metrics.js and update re-exports and references.
