+++ src/shared/tests/metrics-utils.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../../plugin/src/reporting/metrics.js";
+import { createMetricsTracker } from "../reporting.js";
 
 function getCacheKeys(tracker, cacheName = "example") {
     tracker.recordCacheHit(cacheName);
