+++ src/shared/tests/reporting-metrics.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../../plugin/src/reporting/metrics.js";
+import { createMetricsTracker } from "../reporting.js";
 
 test("snapshot exposes accumulated metrics as plain objects", () => {
     const tracker = createMetricsTracker({ category: "demo" });
